requirements.txt
amqp==5.1.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
aspose-words==22.4.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
celery==5.2.6
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.5
django-4-jet==1.0.8
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
eventlet==0.33.0
fasteners==0.14.1
filelock==3.0.12
fire==0.4.0
fonttools==4.33.3
future==0.18.2
google-api-python-client==1.4.1
greenlet==1.1.2
gunicorn==20.1.0
htmldocx==0.0.6
httpie==1.0.3
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mongoengine==0.24.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
packaging==21.3
paramiko==2.6.0
pdf2docx==0.5.3
pdfkit==1.0.0
pexpect==4.6.0
Pillow==7.0.0
prompt-toolkit==3.0.29
protobuf==3.6.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.1.1
PyMuPDF==1.19.6
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-docx==0.8.11
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.2.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
systemd-python==234
termcolor==1.1.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
xkit==0.0.0
zipp==1.0.0
